+++ .gitmodules
@@ -1,5 +1,5 @@
-[submodule "test_util/std"]
-	path = test_util/std
+[submodule "tests/util/std"]
+	path = tests/util/std
 	url = https://github.com/denoland/deno_std
 	shallow = true
 [submodule "test_util/wpt"]
