+++ tests/config/deno.json
@@ -1,6 +1,6 @@
 {
   "imports": {
     "@test_util/": "../../test_util/",
-    "@std/": "../../test_util/std/"
+    "@std/": "../util/std/"
   }
 }
