+++ tests/node_compat/deno.json
@@ -1,6 +1,6 @@
 {
   "imports": {
     "@test_util/": "../../test_util/",
-    "@std/": "../../test_util/std/"
+    "@std/": "../util/std/"
   }
 }
