+++ tests/testdata/cert/listen_tls_alpn_fail.ts
@@ -1,4 +1,4 @@
-import { assertRejects } from "../../../test_util/std/assert/mod.ts";
+import { assertRejects } from "../../../tests/util/std/assert/mod.ts";
 
 const listener = Deno.listenTls({
   port: Number(Deno.args[0]),
