+++ tests/testdata/compile/standalone_follow_redirects_2.js
@@ -2,4 +2,4 @@
 import {
   assertNotEquals as _a,
   assertStrictEquals as _b,
-} from "../../../test_util/std/assert/mod.ts";
+} from "../../../tests/util/std/assert/mod.ts";
