+++ tests/testdata/run/045_proxy_test.ts
@@ -1,6 +1,6 @@
 // Copyright 2018-2024 the Deno authors. All rights reserved. MIT license.
-import { Server } from "../../../test_util/std/http/server.ts";
-import { assertEquals } from "../../../test_util/std/assert/mod.ts";
+import { Server } from "../../../tests/util/std/http/server.ts";
+import { assertEquals } from "../../../tests/util/std/assert/mod.ts";
 
 const addr = Deno.args[1] || "localhost:4555";
