+++ tests/testdata/run/import_meta/main.ts
@@ -1,4 +1,4 @@
-import { assertThrows } from "../../../../test_util/std/assert/mod.ts";
+import { assertThrows } from "../../../../tests/util/std/assert/mod.ts";
 import "http://localhost:4545/run/import_meta/other.ts";
 import "./other.ts";
