+++ tests/testdata/run/onload/imported.ts
@@ -1,5 +1,5 @@
 // deno-lint-ignore-file no-window-prefix
-import { assert } from "../../../../test_util/std/assert/mod.ts";
+import { assert } from "../../../../tests/util/std/assert/mod.ts";
 import "./nest_imported.ts";
 
 const handler = (e: Event) => {
