+++ tests/testdata/run/onload/main.ts
@@ -1,5 +1,5 @@
 // deno-lint-ignore-file no-window-prefix no-prototype-builtins
-import { assert } from "../../../../test_util/std/assert/mod.ts";
+import { assert } from "../../../../tests/util/std/assert/mod.ts";
 import "./imported.ts";
 
 assert(window.hasOwnProperty("onload"));
