+++ tests/testdata/run/onload/nest_imported.ts
@@ -1,5 +1,5 @@
 // deno-lint-ignore-file no-window-prefix
-import { assert } from "../../../../test_util/std/assert/mod.ts";
+import { assert } from "../../../../tests/util/std/assert/mod.ts";
 
 const handler = (e: Event) => {
   assert(e.type === "beforeunload" ? e.cancelable : !e.cancelable);
