+++ tests/testdata/run/tls_starttls.js
@@ -1,6 +1,6 @@
-import { assert, assertEquals } from "../../../test_util/std/assert/mod.ts";
-import { BufReader } from "../../../test_util/std/io/buf_reader.ts";
-import { BufWriter } from "../../../test_util/std/io/buf_writer.ts";
+import { assert, assertEquals } from "../../../tests/util/std/assert/mod.ts";
+import { BufReader } from "../../../tests/util/std/io/buf_reader.ts";
+import { BufWriter } from "../../../tests/util/std/io/buf_writer.ts";
 import { TextProtoReader } from "./textproto.ts";
 
 const encoder = new TextEncoder();
