+++ tests/testdata/test/allow_none.ts
@@ -1,4 +1,4 @@
-import { unreachable } from "../../../test_util/std/assert/mod.ts";
+import { unreachable } from "../../../tests/util/std/assert/mod.ts";
 
 const permissions: Deno.PermissionName[] = [
   "read",
