+++ tests/testdata/workers/deno_worker.ts
@@ -1,4 +1,4 @@
-import { assert } from "../../../test_util/std/assert/mod.ts";
+import { assert } from "../../../tests/util/std/assert/mod.ts";
 
 onmessage = function (e) {
   if (typeof self.Deno === "undefined") {
