+++ tools/node_compat/deno.json
@@ -1,7 +1,7 @@
 {
   "imports": {
     "@test_util/": "../../test_util/",
-    "@std/": "../../test_util/std/"
+    "@std/": "../../tests/util/std/"
   },
   "tasks": {
     "setup": "deno run --allow-read --allow-write ./setup.ts",
