+++ tools/wpt.ts
@@ -32,9 +32,15 @@ import {
   updateManifest,
   wptreport,
 } from "./wpt/utils.ts";
-import { pooledMap } from "../test_util/std/async/pool.ts";
-import { blue, bold, green, red, yellow } from "../test_util/std/fmt/colors.ts";
-import { writeAll, writeAllSync } from "../test_util/std/streams/write_all.ts";
+import { pooledMap } from "../tests/util/std/async/pool.ts";
+import {
+  blue,
+  bold,
+  green,
+  red,
+  yellow,
+} from "../tests/util/std/fmt/colors.ts";
+import { writeAll, writeAllSync } from "../tests/util/std/streams/write_all.ts";
 import { saveExpectation } from "./wpt/utils.ts";
 
 class TestFilter {
